moved setupProxyAgents to until function
