+++ packages/bruno-electron/src/utils/proxy-util.js
@@ -1,6 +1,11 @@
 const parseUrl = require('url').parse;
-const { isEmpty } = require('lodash');
+const https = require('https');
+const { isEmpty, get, isUndefined, isNull } = require('lodash');
 const { HttpsProxyAgent } = require('https-proxy-agent');
+const { interpolateString } = require('../ipc/network/interpolate-string');
+const { SocksProxyAgent } = require('socks-proxy-agent');
+const { HttpProxyAgent } = require('http-proxy-agent');
+const { preferencesUtil } = require('../store/preferences');
 
 const DEFAULT_PORTS = {
   ftp: 21,
@@ -79,7 +84,79 @@ class PatchedHttpsProxyAgent extends HttpsProxyAgent {
   }
 }
 
+function setupProxyAgents({
+  requestConfig,
+  proxyMode,
+  proxyConfig,
+  httpsAgentRequestFields,
+  interpolationOptions
+}) {
+  if (proxyMode === 'on') {
+    const shouldProxy = shouldUseProxy(requestConfig.url, get(proxyConfig, 'bypassProxy', ''));
+    if (shouldProxy) {
+      const proxyProtocol = interpolateString(get(proxyConfig, 'protocol'), interpolationOptions);
+      const proxyHostname = interpolateString(get(proxyConfig, 'hostname'), interpolationOptions);
+      const proxyPort = interpolateString(get(proxyConfig, 'port'), interpolationOptions);
+      const proxyAuthEnabled = get(proxyConfig, 'auth.enabled', false);
+      const socksEnabled = proxyProtocol.includes('socks');
+
+      let uriPort = isUndefined(proxyPort) || isNull(proxyPort) ? '' : `:${proxyPort}`;
+      let proxyUri;
+      if (proxyAuthEnabled) {
+        const proxyAuthUsername = interpolateString(get(proxyConfig, 'auth.username'), interpolationOptions);
+        const proxyAuthPassword = interpolateString(get(proxyConfig, 'auth.password'), interpolationOptions);
+        proxyUri = `${proxyProtocol}://${proxyAuthUsername}:${proxyAuthPassword}@${proxyHostname}${uriPort}`;
+      } else {
+        proxyUri = `${proxyProtocol}://${proxyHostname}${uriPort}`;
+      }
+
+      if (socksEnabled) {
+        requestConfig.httpAgent = new SocksProxyAgent(proxyUri);
+        requestConfig.httpsAgent = new SocksProxyAgent(proxyUri, httpsAgentRequestFields);
+      } else {
+        requestConfig.httpAgent = new HttpProxyAgent(proxyUri);
+        requestConfig.httpsAgent = new PatchedHttpsProxyAgent(
+          proxyUri,
+          Object.keys(httpsAgentRequestFields).length > 0 ? { ...httpsAgentRequestFields } : undefined
+        );
+      }
+    } else {
+      // If proxy should not be used, set default HTTPS agent
+      requestConfig.httpsAgent = new https.Agent(httpsAgentRequestFields);
+    }
+  } else if (proxyMode === 'system') {
+    const { http_proxy, https_proxy, no_proxy } = preferencesUtil.getSystemProxyEnvVariables();
+    const shouldUseSystemProxy = shouldUseProxy(url, no_proxy || '');
+    if (shouldUseSystemProxy) {
+      try {
+        if (http_proxy?.length) {
+          new URL(http_proxy);
+          requestConfig.httpAgent = new HttpProxyAgent(http_proxy);
+        }
+      } catch (error) {
+        throw new Error('Invalid system http_proxy');
+      }
+      try {
+        if (https_proxy?.length) {
+          new URL(https_proxy);
+          requestConfig.httpsAgent = new PatchedHttpsProxyAgent(
+            https_proxy,
+            Object.keys(httpsAgentRequestFields).length > 0 ? { ...httpsAgentRequestFields } : undefined
+          );
+        }
+      } catch (error) {
+        throw new Error('Invalid system https_proxy');
+      }
+    } else {
+      requestConfig.httpsAgent = new https.Agent(httpsAgentRequestFields);
+    }
+  } else if (Object.keys(httpsAgentRequestFields).length > 0) {
+    requestConfig.httpsAgent = new https.Agent(httpsAgentRequestFields);
+  }
+}
+
 module.exports = {
   shouldUseProxy,
-  PatchedHttpsProxyAgent
+  PatchedHttpsProxyAgent,
+  setupProxyAgents
 };
